rehome more tests
